+++ djangocms_references/cms_config.py
@@ -8,7 +8,6 @@ from django.utils.translation import gettext_lazy as _
 
 from cms.app_base import CMSAppConfig, CMSAppExtension
 from cms.plugin_base import CMSPlugin
-from cms.plugin_pool import plugin_pool
 
 from djangocms_alias.models import AliasPlugin
 
@@ -25,11 +24,29 @@ class ReferencesCMSExtension(CMSAppExtension):
     def __init__(self):
         self.reference_models = self._make_default()
         self.reference_plugins = self._make_default()
+        self.reference_complex_relationships = self._make_default()
         self.list_extra_columns = []
         self.list_queryset_modifiers = []
 
     def _make_default(self):
         return defaultdict(lambda: defaultdict(set))
+
+    def get_nested_relationship(self, model, fields):
+        """Throws ImproperlyConfigured if relationship provided doesn't exist,
+        returns the last model in the relationship
+
+        :param model: Target model
+        :param fields: List of fields targeting next model
+        :returns: A related model at the end of the relationship
+        """
+        for validation_field in fields:
+            try:
+                model = getattr(model, validation_field).field.related_model
+            except (ValueError, TypeError) as e:
+                raise ImproperlyConfigured(
+                    "Elements of the reference_fields list should be (model, field_name) tuples"
+                ) from e
+        return model
 
     def register_fields(self, definitions):
         """Registers relations to enable reference retrieval.
@@ -49,16 +66,19 @@ class ReferencesCMSExtension(CMSAppExtension):
                 raise ImproperlyConfigured(
                     "Elements of the reference_fields list should be (model, field_name) tuples"
                 ) from e
-            field = model._meta.get_field(field_name)
-            related_model = field.related_model
+            if "__" in field_name:
+                fields = field_name.split("__")
+                related_model = self.get_nested_relationship(model, fields)
+            else:
+                field = model._meta.get_field(field_name)
+                related_model = field.related_model
             if (
                 issubclass(model, (CMSPlugin,))
-                and model.__name__ in plugin_pool.plugins
             ):
                 store = self.reference_plugins
             else:
                 store = self.reference_models
-            store[related_model][model].add(field.name)
+            store[related_model][model].add(field_name)
 
     def configure_list_extra_columns(self, extra_columns):
         """Registers additional columns to be displayed in the reference