+++ tests/settings.py
@@ -6,6 +6,7 @@ HELPER_SETTINGS = {
         "djangocms_versioning",
         "djangocms_references.test_utils.app_1",
         "djangocms_references.test_utils.polls",
+        "djangocms_references.test_utils.nested_references_app",
     ],
     "MIGRATION_MODULES": {
         "auth": None,