+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 
 Unreleased
 ==========
+* fix: Issue where a plugins name isn't the same as the registered plugin, treating a plugin as an object.
+* feat: Handle references objects that can be deeply nested through model relationships.
 
 1.1.1 (2022-04-11)
 ==================